COZMO-11583 Proper logic fix (#4320)
* COZMO-11583 Proper logic fix
EnableTestStateMessage has been completely removed and the RTIP now manages
logic for sending TestState messages or not based on the complement of
whether it's sending regular state messages or not. This ensures a constant
stream of CLAD messages from the RTIP going through the WiFi so the watchdog
never bites.
* Reset motor calibration on disconnect so that waitForFirstMotorCalibAfterConnect_ behaves properly on reconnect
* Simplify SendTestStateMsg logic even further based on now corrected logic in cozmoBot.cpp
Double send backpack layer command in FactoryTest module startup to make up for issue
where sometimes first message WiFi -> RTIP is dropped.
* Remove bad print statement that was spilling flash memory out debug without a particular termination.
* COZMO-11788 Inserted a delay in I2SPI synchronized callback
Apparently we need to wait for the first few I2SPI buffers to flush out
after synchronization before starting to send CLAD messages.
* Ensure that OTA success message is sent before rebooting.

+++ robot/espressif/app/application/upgradeController.cpp
@@ -172,7 +172,6 @@ namespace UpgradeController {
     const char* versionStr = os_strstr(json, VERSION_TAG);
     if (versionStr) {
       COZMO_VERSION_ID = atoi(versionStr + os_strlen(VERSION_TAG));
-      os_printf("version str '%s' = %d\r\n",versionStr+os_strlen(VERSION_TAG), COZMO_VERSION_ID);
     }
     else COZMO_VERSION_ID = -2;
     static const char* TIME_TAG = "\"time\": ";
@@ -954,9 +953,10 @@ namespace UpgradeController {
         else
         {
           AnkiDebug( 172, "UpgradeController.state", 463, "Reboot", 0);
+          ack.result = SUCCESS;
+          RobotInterface::SendMessage(ack);
           self.phase   = OTAT_Reboot;
-          //self.phase   = OTAT_Apply_RTIP;
-          //self.timer = system_get_time() + 10000000; // 10 second timeout
+          self.timer   = system_get_time() + 2000000;
           self.retries = MAX_RETRIES;
         }
         break;
@@ -1016,10 +1016,11 @@ namespace UpgradeController {
           default: // If the RTIP has booted we'll get other gibberish
           {
             AnkiDebug( 172, "UpgradeController.state", 463, "Reboot", 0);
-            #if DEBUG_OTA
-            os_printf("Done programming, send command done\r\n");
-            #endif
-            self.phase = OTAT_Reboot;
+            ack.result = SUCCESS;
+            RobotInterface::SendMessage(ack);
+            self.phase   = OTAT_Reboot;
+            self.timer   = system_get_time() + 2000000;
+            self.retries = MAX_RETRIES;
             break;
           }
         }
@@ -1027,15 +1028,16 @@ namespace UpgradeController {
       }
       case OTAT_Reboot:
       {
-        i2spiBootloaderCommandDone();
-        ack.result = SUCCESS;
-        RobotInterface::SendMessage(ack);
-        AnkiDebug( 172, "UpgradeController.state", 464, "Wait for reboot", 0);
-        self.phase = OTAT_Wait_For_Reboot;
-        #if DEBUG_OTA
-        os_printf("OTAT Wait for Reboot\r\n");
-        #endif
-        self.timer = system_get_time() + 100000; // delay for messages to clear
+        if (clientQueueFlushed() || system_get_time() > self.timer)
+        {
+          i2spiBootloaderCommandDone();
+          AnkiDebug( 172, "UpgradeController.state", 464, "Wait for reboot", 0);
+          self.phase = OTAT_Wait_For_Reboot;
+          #if DEBUG_OTA
+          os_printf("OTAT Wait for Reboot\r\n");
+          #endif
+          self.timer = system_get_time() + 100000; // delay for messages to clear
+        }
         break;
       }
       case OTAT_Wait_For_Reboot:
